+++ rds-20140815/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-04-14 Version: 15.8.2
+- Update API DescribeRCDisks: add request parameters Status.
+- Update API RunRCInstances: add request parameters BusinessInfo.
+
+
 2026-04-01 Version: 15.8.0
 - Support API ModifyDBInstance.